+++ justify/src/main/java/org/leadpony/justify/internal/schema/io/AbstractJsonSchemaReader.java
@@ -28,6 +28,7 @@ import org.leadpony.justify.api.JsonValidatingException;
 import org.leadpony.justify.api.Problem;
 import org.leadpony.justify.internal.problem.ProblemBuilder;
 import org.leadpony.justify.internal.problem.ProblemBuilderFactory;
+import org.leadpony.justify.internal.problem.ProblemRenderer;
 
 /**
  * A minimal implementation of {@link JsonSchemaReader}.
@@ -113,7 +114,7 @@ abstract class AbstractJsonSchemaReader implements JsonSchemaReader, ProblemBuil
 
     protected void dispatchProblems() {
         if (!problems.isEmpty()) {
-            throw new JsonValidatingException(this.problems);
+            throw new JsonValidatingException(this.problems, ProblemRenderer.DEFAULT_RENDERER);
         }
     }